Update Anomalib data with the new structure
