+++ src/anomalib/data/base/depth.py
@@ -8,8 +8,8 @@ import cv2
 import numpy as np
 import torch
 
+from anomalib.data import TaskType
 from anomalib.data.base.dataset import AnomalibDataset
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import masks_to_boxes, read_depth_image, read_image
