+++ src/anomalib/data/depth/folder_3d.py
@@ -7,14 +7,13 @@ This script creates a custom dataset from a folder.
 # SPDX-License-Identifier: Apache-2.0
 
 
-from collections.abc import Sequence
 from pathlib import Path
 
 import albumentations as A  # noqa: N812
-from pandas import DataFrame
+from pandas import DataFrame, isna
 
-from anomalib.data.base import AnomalibDataModule, AnomalibDataset
-from anomalib.data.task_type import TaskType
+from anomalib.data import TaskType
+from anomalib.data.base import AnomalibDataModule, AnomalibDepthDataset
 from anomalib.data.utils import (
     DirType,
     InputNormalizationMethod,
@@ -27,60 +26,53 @@ from anomalib.data.utils import (
 from anomalib.data.utils.path import _prepare_files_labels, _resolve_path
 
 
-def make_folder_dataset(
-    normal_dir: str | Path | Sequence[str | Path],
+def make_folder3d_dataset(
+    normal_dir: str | Path,
     root: str | Path | None = None,
-    abnormal_dir: str | Path | Sequence[str | Path] | None = None,
-    normal_test_dir: str | Path | Sequence[str | Path] | None = None,
-    mask_dir: str | Path | Sequence[str | Path] | None = None,
+    abnormal_dir: str | Path | None = None,
+    normal_test_dir: str | Path | None = None,
+    mask_dir: str | Path | None = None,
+    normal_depth_dir: str | Path | None = None,
+    abnormal_depth_dir: str | Path | None = None,
+    normal_test_depth_dir: str | Path | None = None,
     split: str | Split | None = None,
     extensions: tuple[str, ...] | None = None,
 ) -> DataFrame:
     """Make Folder Dataset.
 
     Args:
-        normal_dir (str | Path | Sequence): Path to the directory containing normal images.
+        normal_dir (str | Path): Path to the directory containing normal images.
         root (str | Path | None): Path to the root directory of the dataset.
-        abnormal_dir (str | Path | Sequence | None, optional): Path to the directory containing abnormal images.
-        normal_test_dir (str | Path | Sequence | None, optional): Path to the directory containing
+        abnormal_dir (str | Path | None, optional): Path to the directory containing abnormal images.
+        normal_test_dir (str | Path | None, optional): Path to the directory containing
             normal images for the test dataset. Normal test images will be a split of `normal_dir`
             if `None`. Defaults to None.
-        mask_dir (str | Path | Sequence | None, optional): Path to the directory containing
+        mask_dir (str | Path | None, optional): Path to the directory containing
             the mask annotations. Defaults to None.
+        normal_depth_dir (str | Path | None, optional): Path to the directory containing
+            normal depth images for the test dataset. Normal test depth images will be a split of `normal_dir`
+        abnormal_depth_dir (str | Path | None, optional): Path to the directory containing
+            abnormal depth images for the test dataset.
+        normal_test_depth_dir (str | Path | None, optional): Path to the directory containing
+            normal depth images for the test dataset. Normal test images will be a split of `normal_dir`
+            if `None`. Defaults to None.
         split (str | Split | None, optional): Dataset split (ie., Split.FULL, Split.TRAIN or Split.TEST).
             Defaults to None.
         extensions (tuple[str, ...] | None, optional): Type of the image extensions to read from the
             directory.
 
     Returns:
-        DataFrame: an output dataframe containing samples for the requested split (ie., train or test).
+        DataFrame: an output dataframe containing samples for the requested split (ie., train or test)
     """
+    normal_dir = _resolve_path(normal_dir, root)
+    abnormal_dir = _resolve_path(abnormal_dir, root) if abnormal_dir is not None else None
+    normal_test_dir = _resolve_path(normal_test_dir, root) if normal_test_dir is not None else None
+    mask_dir = _resolve_path(mask_dir, root) if mask_dir is not None else None
+    normal_depth_dir = _resolve_path(normal_depth_dir, root) if normal_depth_dir is not None else None
+    abnormal_depth_dir = _resolve_path(abnormal_depth_dir, root) if abnormal_depth_dir is not None else None
+    normal_test_depth_dir = _resolve_path(normal_test_depth_dir, root) if normal_test_depth_dir is not None else None
 
-    def _resolve_path_and_convert_to_list(path: str | Path | Sequence[str | Path] | None) -> list[Path]:
-        """Convert path to list of paths.
-
-        Args:
-            path (str | Path | Sequence | None): Path to replace with Sequence[str | Path].
-
-        Examples:
-            >>> _resolve_path_and_convert_to_list("dir")
-            [Path("path/to/dir")]
-            >>> _resolve_path_and_convert_to_list(["dir1", "dir2"])
-            [Path("path/to/dir1"), Path("path/to/dir2")]
-
-        Returns:
-            list[Path]: The result of path replaced by Sequence[str | Path].
-        """
-        if isinstance(path, Sequence) and not isinstance(path, str):
-            return [_resolve_path(dir_path, root) for dir_path in path]
-        return [_resolve_path(path, root)] if path is not None else []
-
-    # All paths are changed to the List[Path] type and used.
-    normal_dir = _resolve_path_and_convert_to_list(normal_dir)
-    abnormal_dir = _resolve_path_and_convert_to_list(abnormal_dir)
-    normal_test_dir = _resolve_path_and_convert_to_list(normal_test_dir)
-    mask_dir = _resolve_path_and_convert_to_list(mask_dir)
-    assert len(normal_dir) > 0, "A folder location must be provided in normal_dir."
+    assert normal_dir.is_dir(), "A folder location must be provided in normal_dir."
 
     filenames = []
     labels = []
@@ -92,14 +84,22 @@ def make_folder_dataset(
     if normal_test_dir:
         dirs[DirType.NORMAL_TEST] = normal_test_dir
 
+    if normal_depth_dir:
+        dirs[DirType.NORMAL_DEPTH] = normal_depth_dir
+
+    if abnormal_depth_dir:
+        dirs[DirType.ABNORMAL_DEPTH] = abnormal_depth_dir
+
+    if normal_test_depth_dir:
+        dirs[DirType.NORMAL_TEST_DEPTH] = normal_test_depth_dir
+
     if mask_dir:
         dirs[DirType.MASK] = mask_dir
 
-    for dir_type, paths in dirs.items():
-        for path in paths:
-            filename, label = _prepare_files_labels(path, dir_type, extensions)
-            filenames += filename
-            labels += label
+    for dir_type, path in dirs.items():
+        filename, label = _prepare_files_labels(path, dir_type, extensions)
+        filenames += filename
+        labels += label
 
     samples = DataFrame({"image_path": filenames, "label": labels})
     samples = samples.sort_values(by="image_path", ignore_index=True)
@@ -113,22 +113,46 @@ def make_folder_dataset(
     samples.label_index = samples.label_index.astype("Int64")
 
     # If a path to mask is provided, add it to the sample dataframe.
+    if normal_depth_dir is not None:
+        samples.loc[samples.label == DirType.NORMAL, "depth_path"] = samples.loc[
+            samples.label == DirType.NORMAL_DEPTH
+        ].image_path.to_numpy()
+        samples.loc[samples.label == DirType.ABNORMAL, "depth_path"] = samples.loc[
+            samples.label == DirType.ABNORMAL_DEPTH
+        ].image_path.to_numpy()
 
-    if len(mask_dir) > 0 and len(abnormal_dir) > 0:
+        if normal_test_dir is not None:
+            samples.loc[samples.label == DirType.NORMAL_TEST, "depth_path"] = samples.loc[
+                samples.label == DirType.NORMAL_TEST_DEPTH
+            ].image_path.to_numpy()
+
+        # make sure every rgb image has a corresponding depth image and that the file exists
+        assert (
+            samples.loc[samples.label_index == LabelName.ABNORMAL]
+            .apply(lambda x: Path(x.image_path).stem in Path(x.depth_path).stem, axis=1)
+            .all()
+        ), "Mismatch between anomalous images and depth images. Make sure the mask files in 'xyz' \
+            folder follow the same naming convention as the anomalous images in the dataset \
+            (e.g. image: '000.png', depth: '000.tiff')."
+
+        assert samples.depth_path.apply(
+            lambda x: Path(x).exists() if not isna(x) else True,
+        ).all(), "missing depth image files"
+
+        samples = samples.astype({"depth_path": "str"})
+
+    # If a path to mask is provided, add it to the sample dataframe.
+    if mask_dir is not None and abnormal_dir is not None:
         samples.loc[samples.label == DirType.ABNORMAL, "mask_path"] = samples.loc[
             samples.label == DirType.MASK
         ].image_path.to_numpy()
         samples["mask_path"] = samples["mask_path"].fillna("")
         samples = samples.astype({"mask_path": "str"})
 
-        # make sure all every rgb image has a corresponding mask image.
-        assert (
-            samples.loc[samples.label_index == LabelName.ABNORMAL]
-            .apply(lambda x: Path(x.image_path).stem in Path(x.mask_path).stem, axis=1)
-            .all()
-        ), "Mismatch between anomalous images and mask images. Make sure the mask files \
-            folder follow the same naming convention as the anomalous images in the dataset \
-            (e.g. image: '000.png', mask: '000.png')."
+        # make sure all the files exist
+        assert samples.mask_path.apply(
+            lambda x: Path(x).exists() if x != "" else True,
+        ).all(), f"missing mask files, mask_dir={mask_dir}"
     else:
         samples["mask_path"] = ""
 
@@ -155,7 +179,7 @@ def make_folder_dataset(
     return samples
 
 
-class FolderDataset(AnomalibDataset):
+class Folder3DDataset(AnomalibDepthDataset):
     """Folder dataset.
 
     Args:
@@ -163,13 +187,20 @@ class FolderDataset(AnomalibDataset):
         transform (A.Compose): Albumentations Compose object describing the transforms that are applied to the inputs.
         split (str | Split | None): Fixed subset split that follows from folder structure on file system.
             Choose from [Split.FULL, Split.TRAIN, Split.TEST]
-        normal_dir (str | Path | Sequence): Path to the directory containing normal images.
+        normal_dir (str | Path): Path to the directory containing normal images.
         root (str | Path | None): Root folder of the dataset.
-        abnormal_dir (str | Path | Sequence | None, optional): Path to the directory containing abnormal images.
-        normal_test_dir (str | Path | Sequence | None, optional): Path to the directory containing
+        abnormal_dir (str | Path | None, optional): Path to the directory containing abnormal images.
+        normal_test_dir (str | Path | None, optional): Path to the directory containing
             normal images for the test dataset. Defaults to None.
-        mask_dir (str | Path | Sequence | None, optional): Path to the directory containing
+        mask_dir (str | Path | None, optional): Path to the directory containing
             the mask annotations. Defaults to None.
+        normal_depth_dir (str | Path | None, optional): Path to the directory containing
+            normal depth images for the test dataset. Normal test depth images will be a split of `normal_dir`
+        abnormal_depth_dir (str | Path | None, optional): Path to the directory containing
+            abnormal depth images for the test dataset.
+        normal_test_depth_dir (str | Path | None, optional): Path to the directory containing
+            normal depth images for the test dataset. Normal test images will be a split of `normal_dir`
+            if `None`. Defaults to None.
         extensions (tuple[str, ...] | None, optional): Type of the image extensions to read from the
             directory.
         val_split_mode (ValSplitMode): Setting that determines how the validation subset is obtained.
@@ -183,11 +214,14 @@ class FolderDataset(AnomalibDataset):
         self,
         task: TaskType,
         transform: A.Compose,
-        normal_dir: str | Path | Sequence[str | Path],
+        normal_dir: str | Path,
         root: str | Path | None = None,
-        abnormal_dir: str | Path | Sequence[str | Path] | None = None,
-        normal_test_dir: str | Path | Sequence[str | Path] | None = None,
-        mask_dir: str | Path | Sequence[str | Path] | None = None,
+        abnormal_dir: str | Path | None = None,
+        normal_test_dir: str | Path | None = None,
+        mask_dir: str | Path | None = None,
+        normal_depth_dir: str | Path | None = None,
+        abnormal_depth_dir: str | Path | None = None,
+        normal_test_depth_dir: str | Path | None = None,
         split: str | Split | None = None,
         extensions: tuple[str, ...] | None = None,
     ) -> None:
@@ -199,34 +233,47 @@ class FolderDataset(AnomalibDataset):
         self.abnormal_dir = abnormal_dir
         self.normal_test_dir = normal_test_dir
         self.mask_dir = mask_dir
+        self.normal_depth_dir = normal_depth_dir
+        self.abnormal_depth_dir = abnormal_depth_dir
+        self.normal_test_depth_dir = normal_test_depth_dir
         self.extensions = extensions
 
     def _setup(self) -> None:
         """Assign samples."""
-        self.samples = make_folder_dataset(
+        self.samples = make_folder3d_dataset(
             root=self.root,
             normal_dir=self.normal_dir,
             abnormal_dir=self.abnormal_dir,
             normal_test_dir=self.normal_test_dir,
             mask_dir=self.mask_dir,
+            normal_depth_dir=self.normal_depth_dir,
+            abnormal_depth_dir=self.abnormal_depth_dir,
+            normal_test_depth_dir=self.normal_test_depth_dir,
             split=self.split,
             extensions=self.extensions,
         )
 
 
-class Folder(AnomalibDataModule):
+class Folder3D(AnomalibDataModule):
     """Folder DataModule.
 
     Args:
-        normal_dir (str | Path | Sequence): Name of the directory containing normal images.
+        normal_dir (str | Path): Name of the directory containing normal images.
             Defaults to "normal".
         root (str | Path | None): Path to the root folder containing normal and abnormal dirs.
-        abnormal_dir (str | Path | None | Sequence): Name of the directory containing abnormal images.
+        abnormal_dir (str | Path | None): Name of the directory containing abnormal images.
             Defaults to "abnormal".
-        normal_test_dir (str | Path | Sequence | None, optional): Path to the directory containing
+        normal_test_dir (str | Path | None, optional): Path to the directory containing
             normal images for the test dataset. Defaults to None.
-        mask_dir (str | Path | Sequence | None, optional): Path to the directory containing
+        mask_dir (str | Path | None, optional): Path to the directory containing
             the mask annotations. Defaults to None.
+        normal_depth_dir (str | Path | None, optional): Path to the directory containing
+            normal depth images for the test dataset. Normal test depth images will be a split of `normal_dir`
+        abnormal_depth_dir (str | Path | None, optional): Path to the directory containing
+            abnormal depth images for the test dataset.
+        normal_test_depth_dir (str | Path | None, optional): Path to the directory containing
+            normal depth images for the test dataset. Normal test images will be a split of `normal_dir`
+            if `None`. Defaults to None.
         normal_split_ratio (float, optional): Ratio to split normal training images and add to the
             test set in case test set doesn't contain any normal images.
             Defaults to 0.2.
@@ -257,12 +304,14 @@ class Folder(AnomalibDataModule):
 
     def __init__(
         self,
-        normal_dir: str | Path | Sequence[str | Path],
+        normal_dir: str | Path,
         root: str | Path | None = None,
-        abnormal_dir: str | Path | Sequence[str | Path] | None = None,
-        normal_test_dir: str | Path | Sequence[str | Path] | None = None,
-        mask_dir: str | Path | Sequence[str | Path] | None = None,
-        normal_split_ratio: float = 0.2,
+        abnormal_dir: str | Path | None = None,
+        normal_test_dir: str | Path | None = None,
+        mask_dir: str | Path | None = None,
+        normal_depth_dir: str | Path | None = None,
+        abnormal_depth_dir: str | Path | None = None,
+        normal_test_depth_dir: str | Path | None = None,
         extensions: tuple[str] | None = None,
         image_size: int | tuple[int, int] = (256, 256),
         center_crop: int | tuple[int, int] | None = None,
@@ -290,16 +339,6 @@ class Folder(AnomalibDataModule):
             seed=seed,
         )
 
-        if task == TaskType.SEGMENTATION and test_split_mode == TestSplitMode.FROM_DIR and mask_dir is None:
-            msg = (
-                f"Segmentation task requires mask directory if test_split_mode is {test_split_mode}. "
-                "You could set test_split_mode to {TestSplitMode.NONE} or provide a mask directory."
-            )
-            raise ValueError(
-                msg,
-            )
-
-        self.normal_split_ratio = normal_split_ratio
         transform_train = get_transforms(
             config=transform_config_train,
             image_size=image_size,
@@ -313,7 +352,7 @@ class Folder(AnomalibDataModule):
             normalization=InputNormalizationMethod(normalization),
         )
 
-        self.train_data = FolderDataset(
+        self.train_data = Folder3DDataset(
             task=task,
             transform=transform_train,
             split=Split.TRAIN,
@@ -322,10 +361,13 @@ class Folder(AnomalibDataModule):
             abnormal_dir=abnormal_dir,
             normal_test_dir=normal_test_dir,
             mask_dir=mask_dir,
+            normal_depth_dir=normal_depth_dir,
+            abnormal_depth_dir=abnormal_depth_dir,
+            normal_test_depth_dir=normal_test_depth_dir,
             extensions=extensions,
         )
 
-        self.test_data = FolderDataset(
+        self.test_data = Folder3DDataset(
             task=task,
             transform=transform_eval,
             split=Split.TEST,
@@ -333,6 +375,9 @@ class Folder(AnomalibDataModule):
             normal_dir=normal_dir,
             abnormal_dir=abnormal_dir,
             normal_test_dir=normal_test_dir,
+            normal_depth_dir=normal_depth_dir,
+            abnormal_depth_dir=abnormal_depth_dir,
+            normal_test_depth_dir=normal_test_depth_dir,
             mask_dir=mask_dir,
             extensions=extensions,
         )
