+++ src/anomalib/data/image/btech.py
@@ -20,8 +20,8 @@ import pandas as pd
 from pandas.core.frame import DataFrame
 from tqdm import tqdm
 
+from anomalib.data import TaskType
 from anomalib.data.base import AnomalibDataModule, AnomalibDataset
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
