+++ src/anomalib/data/utils/task_type.py
