+++ src/anomalib/data/video/avenue.py
@@ -26,9 +26,9 @@ import numpy as np
 import scipy.io
 from pandas import DataFrame
 
+from anomalib.data import TaskType
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
