+++ src/anomalib/data/video/shanghaitech.py
@@ -25,9 +25,9 @@ import pandas as pd
 import torch
 from pandas import DataFrame
 
+from anomalib.data import TaskType
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
