+++ src/anomalib/data/video/ucsd_ped.py
@@ -15,9 +15,9 @@ import numpy as np
 import torch
 from pandas import DataFrame
 
+from anomalib.data import TaskType
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
-from anomalib.data.task_type import TaskType
 from anomalib.data.utils import (
     DownloadInfo,
     InputNormalizationMethod,
